auth and logic

+++ src/main/java/fontys/magiccardgame/persistence/CardRepository.java
@@ -3,6 +3,6 @@ package fontys.magiccardgame.persistence;
 import fontys.magiccardgame.persistence.entity.CardEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CardRepository extends JpaRepository<CardEntity,Integer>{
+public interface CardRepository extends JpaRepository<CardEntity,Long>{
 
 }
